Rename database columns and enforce model relation graphs
For IdentityProvider and Permission, it makes more sense to have the
column names be idp and permCode instead of the ambiguous "code". By
changing this to idp and permCode, this will allow for more direct one to
one mapping with the API body objects and the underlying database model.
As well, all string field max lengths are now explicitly defined.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>