Refactor and unit test hasPermission authorization middleware
The middleware was too cognitively complex to reason through, so a refactor
to make the boolean logic mostly sequential was done. This makes it easier
to reason through while still preserving the ability to short circuit as
necessary. Unit tests do not cover the complete cartesian product space,
but instead cover most reasonable and reachable edge cases.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>