Expose S3 expiresIn parameter to readObject operations
In the event the user specifies a custom expiresIn query parameter, the
endpoint will generate a direct S3 link with the specified expiresIn, even
if the download query parameter is also specified. This is because if
expiresIn is specified, it does not make sense for COMS to handle the
file transfer directly as it will not be able to temporally restrict access
outside of JWT expirations.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>