Implement searchUsers controller and service
This endpoint will allow users to search the users that COMS is tracking
and completes the minimum round trip necessary for granting permissions
for objects to specific users. This endpoint is considered a discovery
endpoint, so it must by definition be able to support almost any kind of
query shape thrown at it.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>