Ensure searchObject endpoints to require a database
At this time, this endpoint is only useful if there is a database to query
on. Otherwise, it will just return an empty list. If we run COMS in noDB
mode, we should consider leveraging the S3 ListObjectsCommand operation
to directly ask S3 what objects are available at the filePath, as well as
consider zipper merging the results from both S3 and database entries
together.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>