Bugfix user login update differential logic
When we get identityIds with uuids in non-dashed format, it trips our
differential object detector, even though they logically mean the same
thing. This results in unnecessary database updates to the user entry on
every network request, when it should only read at most. To remedy this,
we need to ensure that we pre-process and format the incoming uuid to be
in the canonical lowercase dashed format so that they actually match as
originally intended.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>