Add soft-uuid correction support to query param handlers
As an extension of the previous commit, we should attempt to sanitize
incoming uuid requests if possible, even if they are uppercased or lack
dashes as those uuid string representations are functionally equivalent.
This also moves the JSDoc that was written incorrectly in the controller
layer to the appropriate service layer.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>