Change audit table originalData and newData to json type
Considering the use cases of this audit table, we are not expecting to
ever require the btree indexing capabilities for the json objects recorded.
We expect most scenarios using this table to be one-off explorations
for incident tracking and investigations. As such, the indexing space cost
is not worth it.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>