Block unauthenticated searchObjects requests while not in NoAuth mode
A minor oversight with implementation meant that unauthorized users could
get a list of every object in the system by not authenticating. This
should only be allowed if the application is running in NoAuth mode. We
should consider if this snippet should be a middleware function down the
line should other endpoints also face a similar situation.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>