Modify CORS to explicitly expose certain S3 Amazon headers
WIthout the Access-Control-Expose-Headers header defined, certain browser
calls will not be able to acquire S3 specific headers. This change attempts
to address that by globally defining a list of headers we always want to
be visible to clients. We may need to revisit this when we get to metadata
as this may not be a tenable long-term arbitrary solution.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>