Bugfix Keycloak config crash in non-OIDC Auth modes
There is an unintended bug where COMS chokes with startup where it is
expecting the keycloak.clientId configuration. The expected behavior is
to not look for that config unless keycloak.enabled is true first. We can
ensure this pre-check behavior by delaying the import of the keycloak
component and having it only execute within the appropriate conditionals.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>