Bugfix Database config crash in noDB modes
There is an unintended bug where COMS chokes with startup where it is
expecting database configurations. The expected behavior is to not look
for that config unless db.enabled is true first. We can ensure this
pre-check behavior by delaying the import of the knexfile component and
having it only execute within the appropriate conditionals.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>