Implement utils.isTruthy for more robust truthiness checks
Query parameters can come in a variety of shapes and forms, so we need a
way to reliably determine if a value is semantically true or false. The
isTruthy function does that by explicitly checking for most common "true"
cases and returning a boolean result. This has been retrofitted to all the
controllers that have active and public query parameters.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>