Refactor OpenAPI schemas to have a common nomenclature
There were too many duplications and domain overlap between all of the
schemas. This change forces the adoption of the <domain>-<schema> naming
pattern in order to help cluster and logically organize what each schema
is for, and where to expect it to appear in the API. This should help ease
API definition extensions in the future.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>