Allow isTruthy to passthrough undefined
While ideally we would like isTruthy to be a pure predicate function, in
practice, we want to take advantage of JSON object collapsing which only
happens when an attribute is undefined (effectively false). This is
important because our search filters do not run if the parameter is
undefined, versus when it is explicitly false, as one indicates an explicit
intent to search a condition, whereas the other implies a lack of
importance in the query.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>