Implement Github Actions Pull Request workflows
This change introduces the ability to deploy via Helm to the PR
environment on certain open pull request triggers. It also introduces
the ability to clean up the same deployments once a PR has been closed.
Comments such as the release location as well as the code coverage have
also been included for continuity.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>