Refactor getBucket to be a utility function
Since the dynamic cascade logic for determining bucket information is
highly contextual, we want to make sure this logic does not get duplicated
in multiple places. To this end, this new utility function ensures that
default config information is always returned as long as throwable is false
and the database bucket lookup failed.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>