Add a section in github issue templates requesting for version number
As we begin rolling out to more audiences, we need to ensure that we are
checking and testing on the right versions of the software. Hopefully by
asking for the version number we will be able to ensure that we do not
burn time tracking down bugs/issues which may have already been resolved
in a later build.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>