Refactor service layer tests to handle multiple model sources
Unfortunately due to Jest mock hoisting behavior, we are not able to write
a mock Model once and use it multiple places. Our solution now depends
on defining the methods we expect to see up front and verbose in the Jest
factory function, and depends on us using a new resetReturnThis utility
function to approximately mock the chaining behavior that Objection uses.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>