Apply default Access-Control-Max-Age header for 10 minutes
As most of our expected network operations will be done by browser, there
will be a large volume of preflight options calls dispatched to COMS.
Without the ACMA header, we would be seeing nearly double the network
call volume. ACMA allows us to lower the amount of those preflight calls
by telling the browser it is ok to cache the results.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>