Ensure metadata pruning happens at the end of the association operation
There was a bug where on non-versioned buckets, updating objects would
fail to work because the pruning operation was happening before the new
version has associated to the new metadata records. As a result, the new
metadata values get swept away prematurely, so when the association
happens, it fails because the metadata no longer exists. This commit fixes
that order of operations bug.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>