Add ListObjectV2Command and drop maxkeys constant default
The maximum integer approach to maxkeys does not actually work. In the
event S3 has more than 1000 records, it will cap at 1000 even if requested
otherwise. We drop this constant as it is no longer of use.
ListObjectV2Command is added as it is meant to be used over the original
ListObjectCommand. The original associated function has been marked as
deprecated.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>