Remove coms metadata enforcement from storage service
As we are likely enforcing business level coms-id tag in controller layer,
doing this again in service layer is redundant. We also streamline the
multipart upload command by embedding the Tagging command into the
constructor instead of bolting it on after the fact. This reduces the
number of network calls done to S3 by one while retaining identical
functionality.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>