Update Authentication approach and jsonwebtoken dependency
The jsonwebtoken library introduced a difficult to resolve fault with unit
testing as of 9.0.0 where the decode function can no longer be spied or
mocked. As the decode function is only really necessary if the publicKey
is not specified, we instead reuse isValid's value when it is possible to
do so instead.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>