Update Authentication approach and jsonwebtoken dependency
The jsonwebtoken library introduced a difficult to resolve fault with unit
testing as of 9.0.0 where the decode function can no longer be spied or
mocked. To go around this problem, we switch to using the jwt.verify
method instead, requesting that it verifies on no algorithms instead to
perform similar decode-only behavior.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>