Update S3 client defaults to no longer require explicit apiVersion
The apiVersion is defined by the library by default, so we do not need to
manually define it anymore. We also add in silly and debug level logging
support from the S3 library in order to get another source of data when
we need to track down S3 related issues.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>