Ensure endpoint values are RFC 3986 compliant and do not have trailing '/'
In order to ensure that we force endpoint collision detection to happen,
the endpoints should be standardized in our database. While url and url/
are semantically equal, the database uniqueness constraint does not see
them as the same string. We mitigate this by forcing all endpoints to lack
the trailing slash going forwards.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>