Ensure bucket keys have standardized slash prefix behavior
In general, COMS handles key paths as relative directories. Since object
paths use relative syntax (without a leading slash), we standardize the
key to also follow that precedent. In all cases, folder directories such
as 'foo' and '/foo' need resolve to the same effective directory so that
in the event both are registered in, we do not have duplicate buckets
effectively representing the same bucket location.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>