Consolidate dataConnection to properly bind knex to objection
Previously, we were only binding knex to objection on the checkAll method
once. If it does not get bound then, it fails to bind later when knex does
have a valid connection pool set up. We fix this by making the knex setter
always rebind the incoming knex object to the objection model.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>