Streamline app to gracefully recover from database disruptions
Previously, our application would abort if the database connections
went down. As knex has the ability to restore the db pool already, we
instead ensure that the application sets itself to a non-ready state and
503 reject all incoming network traffic until the database connections are
restored. This ensures that our queue poller and general operations can
function and pause appropriately given the db state. We also introduce a
new 'SERVER_HARDRESET' environment variable which when defined preserves
the older abort behavior and ensure that top level unhandled errors
gracefully shut down the application.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>