Refactor enqueue job to handle one or many job insertions at once
As we are anticipating that we may be inserting many jobs at once, it is
more efficient to batch the insert operations together in one shot. This
refactor allows for optional batching while still ensuring that there is
some degree of value enforcement. Enqueue now returns a number representing
the number of records that were inserted into the queue instead of a
returning a boolean.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>