Implement statusSync and statusDefault endpoint support
In order to reduce code duplication, we reuse the syncBucket controller
with a few value tweaks so that it can handle both endpoint types. We also
move the sync* controller functions to the sync controller in order to
reduce controller pollution of the large object and bucket controllers, and
invoke these functions directly through the router instead. Lastly, some
database layer modifications are done in order to allow bucketId to accept
null fields.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>