Implement _deriveObjectId utility synchronize method
This method focuses on attempting to acquire a previously known 'coms-id'
tag. If it can procedurally find one in the version history, it will yield
that uuid; otherwise a new one will be generated. This is expected to only
be used when creating new COMS Object DB records. A 'coms-id' will be
written back to S3 in the event this is newly assigned.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>