Bugfix isAtPath to accept identical prefix/path values
The previous fix checking for length being less than 2 was not accurate
because it was letting paths with just a trailing slash to pass through.
We roll it back to checking if the length is exactly 1 in difference, and
also do an earlier check for prefix and path being identical in order to
better handle the exception cases.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>