Modify shutdown/cleanup ordering to halt queue processing earlier
Previously the terminate signal would come in, but the queue would continue
working for another 3 seconds or so before halting itself. We change it up
so that when the signal comes in, it immediately tells the queue to stop
before doing the 3 second grace period to kill the database connection.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>