Unit test and refactor queueManager close to not recursively spinlock
There are situations where the spinlock causes a stack overflow instead of
properly waiting for the isBusy state to change. We deal with this by
inverting the control flow to insert in a stored callback reference and
then invoking the callback when isBusy is assigned a false value. This
approach removes the need to poll frequently and instead defers it to the
setter to do the work at the right time.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>