Refactor Problem handling to throw instead of sending
In order to standardize how we deal with 400 and 500 class HTTP responses,
we now always throw a Problem error instead of short circuiting with the
send method. Instead, we now depend on the top level app error handler to
capture and render the Problem. This ensures that we only have one pattern
for problem emission. Other minor various unit tests are also updated.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>