Implement lastModifiedDate tracking support
This change begins to track the LastModified attribute in S3 versions.
However, due to the HeadObjectCommand returning a 404/405 when asking about
a delete marker, we will not be able to back-populate delete marker
version records on the fly. The synchronization operation is not affected
by this shortcoming as it uses a different pathway to surface this data.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>